add locate and locate! methods to MadID to make raising errors optional
this changes the behaviour of locate and adds a locate! method.
the locate! method raises a KeyError error if no class is found in the
registry and an ActiveRecord::RecordNotFoundError if no record is found

the locate method simply ignores those errors and returns nil. This is
helpful when you expect the search for strange IDs but do not care if
nothing is found.

NOTE: this changes the current API