make finders always accessible and default to id column
when MadID is included in a class we extend that class with the finder
methods and by default search in the id column.
so we can always use the find_by_mad_id finder method. This is
especially helpful if you have mad and not mad classes.

also this makes the column configurable with the column option on the
identify_with method.

identify_with :dog, column: 'mad_id'